# Conflicts:
#	src/sources/file.rs
#	website/cue/reference/components/sources/base/file.cue

+++ src/sources/logstash.rs
@@ -38,7 +38,7 @@ use crate::{
 #[configurable_component(source("logstash"))]
 #[derive(Clone, Debug)]
 pub struct LogstashConfig {
-    /// The address to listen for connections on.
+    /// The socket address to listen for connections on.
     address: SocketListenAddr,
 
     #[configurable(derived)]
@@ -610,7 +610,7 @@ impl Decoder for LogstashDecoder {
                     let payload_size = rest.get_u32() as usize;
 
                     if rest.remaining() < payload_size {
-                        src.reserve(payload_size as usize);
+                        src.reserve(payload_size);
                         return Ok(None);
                     }
